ipa: ipahal: ipahal_hw_stats: Fix -Wpointer-to-int-cast
drivers/platform/msm/ipa/ipa_v3/ipahal/ipahal_hw_stats.c:228:12: error: cast to smaller integer type 'int' from 'void *' [-Werror,-Wvoid-pointer-to-int-cast]
        int num = (int)(params);
                  ^~~~~~~~~~~~~

Fix `int num = (int)(params)` by casting pointer value into long,
then compiler assignment will truncate to 32bit to fit `int num` later.

Bug: 171827315

+++ drivers/platform/msm/ipa/ipa_v3/ipahal/ipahal_hw_stats.c
@@ -225,7 +225,7 @@ static struct ipahal_stats_init_pyld *ipahal_generate_init_pyld_flt_rt_v4_5(
 	void *params, bool is_atomic_ctx)
 {
 	struct ipahal_stats_init_pyld *pyld;
-	int num = (int)(params);
+	long num = (long)(params);
 
 	if (num > IPA_MAX_FLT_RT_CNT_INDEX ||
 		num <= 0) {
